# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DiagnosticTestCategory = typeof DiagnosticTestCategory[keyof typeof DiagnosticTestCategory];


export const DiagnosticTestCategory = {
  blood: 'blood',
  imaging: 'imaging',
  packages: 'packages',
} as const;

export interface DiagnosticTest {
  id: string;
  name: string;
  category: DiagnosticTestCategory;
  price: number;
  /** @nullable */
  mrp?: number | null;
  turnaround: string;
  /** @nullable */
  description?: string | null;
  fastingRequired?: boolean;
  popular?: boolean;
}

export interface HealthPackage {
  id: string;
  name: string;
  shortDescription?: string;
  description: string;
  price: number;
  mrp: number;
  parameterCount: number;
  category: string;
  fastingRequired?: boolean;
  includes?: string[];
  /**
     * e.g. "Most Popular", "Best Value"
     * @nullable
     */
  badge?: string | null;
  popular?: boolean;
}

export interface TestSearchResult {
  tests: DiagnosticTest[];
  packages: HealthPackage[];
}

export type CartItemItemType = typeof CartItemItemType[keyof typeof CartItemItemType];


export const CartItemItemType = {
  test: 'test',
  package: 'package',
} as const;

export interface CartItem {
  itemId: string;
  itemType: CartItemItemType;
  name: string;
  price: number;
  quantity: number;
}

export type BookingInputPatientGender = typeof BookingInputPatientGender[keyof typeof BookingInputPatientGender];


export const BookingInputPatientGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export type BookingInputCollectionType = typeof BookingInputCollectionType[keyof typeof BookingInputCollectionType];


export const BookingInputCollectionType = {
  home: 'home',
  walkin: 'walkin',
} as const;

export interface BookingInput {
  patientName: string;
  patientAge: number;
  patientGender: BookingInputPatientGender;
  phone: string;
  address: string;
  pincode: string;
  preferredDate: string;
  preferredTimeSlot: string;
  collectionType: BookingInputCollectionType;
  items: CartItem[];
  totalAmount?: number;
  homeCollectionFee?: number;
  /** @nullable */
  notes?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  sample_collected: 'sample_collected',
  processing: 'processing',
  report_ready: 'report_ready',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  bookingId: string;
  patientName: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  preferredTimeSlot?: string | null;
  collectionType?: string;
  status: BookingStatus;
  totalAmount: number;
  homeCollectionFee?: number;
  /** @nullable */
  reportUrl?: string | null;
  items: CartItem[];
  createdAt: string;
}

export interface PrescriptionInput {
  name: string;
  phone: string;
  preferredArea: string;
  preferredTimeSlot: string;
  fileNames?: string[];
  /** @nullable */
  notes?: string | null;
}

export interface PrescriptionSubmission {
  id: number;
  referenceId: string;
  message: string;
  createdAt: string;
}

export interface Location {
  id: string;
  name: string;
  area: string;
  address: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  hours: string;
  mapsUrl?: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
}

export interface Testimonial {
  id: string;
  name: string;
  rating: number;
  text: string;
  service: string;
  date: string;
  area?: string;
}

export interface SiteStats {
  testsCount: number;
  locationsCount: number;
  yearsExperience: number;
  patientsServed: number;
  turnaroundHours: number;
}

export type SearchTestsParams = {
/**
 * Search query (test name, health concern)
 */
q?: string;
/**
 * Filter by category
 */
category?: SearchTestsCategory;
};

export type SearchTestsCategory = typeof SearchTestsCategory[keyof typeof SearchTestsCategory];


export const SearchTestsCategory = {
  all: 'all',
  blood: 'blood',
  imaging: 'imaging',
  packages: 'packages',
} as const;

export type LookupBookingParams = {
mobile?: string;
bookingId?: string;
};