# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { CartItem } from './cartItem';

export interface Booking {
  id: number;
  bookingId: string;
  patientName: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  preferredTimeSlot?: string | null;
  collectionType?: string;
  status: BookingStatus;
  totalAmount: number;
  homeCollectionFee?: number;
  /** @nullable */
  reportUrl?: string | null;
  items: CartItem[];
  createdAt: Date;
}