# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputCollectionType } from './bookingInputCollectionType';
import type { BookingInputPatientGender } from './bookingInputPatientGender';
import type { CartItem } from './cartItem';

export interface BookingInput {
  patientName: string;
  patientAge: number;
  patientGender: BookingInputPatientGender;
  phone: string;
  address: string;
  pincode: string;
  preferredDate: Date;
  preferredTimeSlot: string;
  collectionType: BookingInputCollectionType;
  items: CartItem[];
  totalAmount?: number;
  homeCollectionFee?: number;
  /** @nullable */
  notes?: string | null;
}