# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItemItemType } from './cartItemItemType';

export interface CartItem {
  itemId: string;
  itemType: CartItemItemType;
  name: string;
  price: number;
  quantity: number;
}