# lib/api-zod/src/generated/types/diagnosticTest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosticTestCategory } from './diagnosticTestCategory';

export interface DiagnosticTest {
  id: string;
  name: string;
  category: DiagnosticTestCategory;
  price: number;
  /** @nullable */
  mrp?: number | null;
  turnaround: string;
  /** @nullable */
  description?: string | null;
  fastingRequired?: boolean;
  popular?: boolean;
}