# lib/api-zod/src/generated/types/healthPackage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthPackage {
  id: string;
  name: string;
  shortDescription?: string;
  description: string;
  price: number;
  mrp: number;
  parameterCount: number;
  category: string;
  fastingRequired?: boolean;
  includes?: string[];
  /**
     * e.g. "Most Popular", "Best Value"
     * @nullable
     */
  badge?: string | null;
  popular?: boolean;
}