# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  id: string;
  name: string;
  area: string;
  address: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  hours: string;
  mapsUrl?: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
}