# lib/api-zod/src/generated/types/prescriptionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */

export interface PrescriptionInput {
  name: string;
  phone: string;
  preferredArea: string;
  preferredTimeSlot: string;
  fileNames?: string[];
  /** @nullable */
  notes?: string | null;
}