# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  testsCount: number;
  locationsCount: number;
  yearsExperience: number;
  patientsServed: number;
  turnaroundHours: number;
}