# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kloud Diagnostics & Imaging API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: string;
  name: string;
  rating: number;
  text: string;
  service: string;
  date: string;
  area?: string;
}